+++ sdk/openai/Azure.AI.OpenAI/src/Custom/CompletionsOptions.cs
@@ -5,7 +5,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.Linq;
 using Azure.Core;