+++ sdk/storage/Azure.Storage.DataMovement/src/CheckpointerExtensions.cs
@@ -2,37 +2,41 @@
 // Licensed under the MIT License.
 
 using System;
+using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
+using Azure.Storage.DataMovement.JobPlan;
 
 namespace Azure.Storage.DataMovement
 {
     internal partial class CheckpointerExtensions
     {
+        internal static async Task<DataTransferStatus> GetJobStatusAsync(
+            this TransferCheckpointer checkpointer,
+            string transferId,
+            CancellationToken cancellationToken = default)
+        {
+            using (Stream stream = await checkpointer.ReadJobPlanFileAsync(
+                transferId,
+                DataMovementConstants.JobPlanFile.JobStatusIndex,
+                DataMovementConstants.IntSizeInBytes,
+                cancellationToken).ConfigureAwait(false))
+            {
+                BinaryReader reader = new BinaryReader(stream);
+                JobPlanStatus jobPlanStatus = (JobPlanStatus)reader.ReadInt32();
+                return jobPlanStatus.ToDataTransferStatus();
+            }
+        }
+
         internal static async Task<bool> IsResumableAsync(
             this TransferCheckpointer checkpointer,
             string transferId,
             CancellationToken cancellationToken)
         {
-            DataTransferState transferState = (DataTransferState) await checkpointer.GetByteValue(
-                transferId,
-                DataMovementConstants.JobPartPlanFile.AtomicJobStatusStateIndex,
-                cancellationToken).ConfigureAwait(false);
-
-            byte hasFailedItemsByte = await checkpointer.GetByteValue(
-                transferId,
-                DataMovementConstants.JobPartPlanFile.AtomicJobStatusHasFailedIndex,
-                cancellationToken).ConfigureAwait(false);
-            bool hasFailedItems = Convert.ToBoolean(hasFailedItemsByte);
-
-            byte hasSkippedItemsByte = await checkpointer.GetByteValue(
-                transferId,
-                DataMovementConstants.JobPartPlanFile.AtomicJobStatusHasSkippedIndex,
-                cancellationToken).ConfigureAwait(false);
-            bool hasSkippedItems = Convert.ToBoolean(hasSkippedItemsByte);
+            DataTransferStatus jobStatus = await checkpointer.GetJobStatusAsync(transferId, cancellationToken).ConfigureAwait(false);
 
             // Transfers marked as fully completed are not resumable
-            return transferState != DataTransferState.Completed || hasFailedItems || hasSkippedItems;
+            return jobStatus.State != DataTransferState.Completed || jobStatus.HasFailedItems || jobStatus.HasSkippedItems;
         }
 
         internal static async Task<DataTransferProperties> GetDataTransferPropertiesAsync(
@@ -60,6 +64,36 @@ namespace Azure.Storage.DataMovement
                 DestinationPath = destPath,
                 IsContainer = isContainer,
             };
+        }
+
+        internal static async Task<bool> IsEnumerationCompleteAsync(
+            this TransferCheckpointer checkpointer,
+            string transferId,
+            CancellationToken cancellationToken)
+        {
+            using (Stream stream = await checkpointer.ReadJobPlanFileAsync(
+                transferId,
+                DataMovementConstants.JobPlanFile.EnumerationCompleteIndex,
+                DataMovementConstants.OneByte,
+                cancellationToken).ConfigureAwait(false))
+            {
+                return Convert.ToBoolean(stream.ReadByte());
+            }
+        }
+
+        internal static async Task OnEnumerationCompleteAsync(
+            this TransferCheckpointer checkpointer,
+            string transferId,
+            CancellationToken cancellationToken)
+        {
+            byte[] enumerationComplete = { Convert.ToByte(true) };
+            await checkpointer.WriteToJobPlanFileAsync(
+                transferId,
+                DataMovementConstants.JobPlanFile.EnumerationCompleteIndex,
+                enumerationComplete,
+                bufferOffset: 0,
+                length: 1,
+                cancellationToken).ConfigureAwait(false);
         }
     }
 }